fix(codegen): DEPYLER-1135 Remove unnecessary numeric casts in push operations
The generate_typed_push function was unconditionally casting integers to i64
and floats to f64. This caused E0308 type mismatch errors when pushing to
Vec<i32> because Python's `int` maps to i32 by default, not i64.

Fix:
- Removed `as i64` cast for Type::Int - trust the type system
- Removed `as f64` cast for Type::Float - trust the type system
- If there's a type mismatch, it should be fixed at the source (type inference)
  not papered over with casts

Also fixes missing struct fields (module_aliases, alias) in test files that
were added in DEPYLER-1136-1139 but not updated in all test locations.

Results:
- Filtered corpus (non-test files): 32.6% → 37.2% (+2 files compiling)
- All 11,309 tests pass
- functional_programming_combined.py now compiles (was failing with E0308)

+++ crates/depyler-core/src/rust_gen/expr_gen_instance_methods.rs
@@ -430,13 +430,18 @@ impl<'a, 'b> ExpressionConverter<'a, 'b> {
                     Ok(parse_quote! { #object_expr.push(#arg) })
                 }
             }
-            // For Vec<i64>, ensure integer type
+            // DEPYLER-1135: For Vec<int>, push directly without cast
+            // The argument should already be typed correctly (i32 by default, or whatever
+            // width was inferred). Casting to i64 causes E0308 when the Vec is Vec<i32>.
+            // Trust the type system - if there's a mismatch, it's a type inference issue
+            // that should be fixed at the source, not papered over with casts.
             Type::Int => {
-                Ok(parse_quote! { #object_expr.push(#arg as i64) })
+                Ok(parse_quote! { #object_expr.push(#arg) })
             }
-            // For Vec<f64>, ensure float type
+            // DEPYLER-1135: For Vec<f64>, push directly without cast
+            // Same reasoning - trust the type system
             Type::Float => {
-                Ok(parse_quote! { #object_expr.push(#arg as f64) })
+                Ok(parse_quote! { #object_expr.push(#arg) })
             }
             // For Vec<bool>, direct push
             Type::Bool => {
